Fix/unpushed changes (#279)
* refactor: refactored types and constants in AudioArray

* refactor: refactored types and constants in AnalyserNode

* refactor: refactored types and constants in AudioBus

* fix: added necessary include

* refactor: refactored types and constants in AudioBuffer

* refactor: refactored types and constants in AudioBufferSourceNode

* refactor: refactored types and constants in BaseAudioContext

* refactor: refactored types and constants in AudioContext

* refactor: refactored types and constants in AudioDecoder and AudioDestinationNode

* refactor: refactored types and constants in AudioNode

* refactor: refactored types and constants in AudioNodeManager

* refactor: refactored types and constants in AudioParam

* refactor: refactored types and constants in AudioScheduledSourceNode

* refactor: refactored types and constants in BiquadFilterNode

* refactor: refactored types and constants in GainNode

* refactor: refactored types and constants in OscillatorNode

* refactor: refactored types and constants in StereoPannerNode

* refactor: refactored types and constants in PeriodicWave

* refactor: refactored types in utils

* ci: yarn format

* fix: fixed analyser types

* refactor: host objects

* fix: fixed analyser node

* refactor: refactored ios audio player

* ci: yarn format

* refactor: replaced M_PI with our PI constant

---------

+++ packages/react-native-audio-api/common/cpp/HostObjects/AnalyserNodeHostObject.h
@@ -126,25 +126,25 @@ class AnalyserNodeHostObject : public AudioNodeHostObject {
 
   JSI_PROPERTY_SETTER(fftSize) {
     auto analyserNode = std::static_pointer_cast<AnalyserNode>(node_);
-    auto fftSize = static_cast<size_t>(value.getNumber());
+    auto fftSize = static_cast<int>(value.getNumber());
     analyserNode->setFftSize(fftSize);
   }
 
   JSI_PROPERTY_SETTER(minDecibels) {
     auto analyserNode = std::static_pointer_cast<AnalyserNode>(node_);
-    auto minDecibels = static_cast<int>(value.getNumber());
+    auto minDecibels = static_cast<float>(value.getNumber());
     analyserNode->setMinDecibels(minDecibels);
   }
 
   JSI_PROPERTY_SETTER(maxDecibels) {
     auto analyserNode = std::static_pointer_cast<AnalyserNode>(node_);
-      auto maxDecibels = static_cast<int>(value.getNumber());
+      auto maxDecibels = static_cast<float>(value.getNumber());
       analyserNode->setMaxDecibels(maxDecibels);
   }
 
   JSI_PROPERTY_SETTER(smoothingTimeConstant) {
     auto analyserNode = std::static_pointer_cast<AnalyserNode>(node_);
-    auto smoothingTimeConstant = static_cast<int>(value.getNumber());
+    auto smoothingTimeConstant = static_cast<float>(value.getNumber());
     analyserNode->setSmoothingTimeConstant(smoothingTimeConstant);
   }
 };
